refactor: extract hero and footer sections
